+++ parity-util-mem/src/allocators.rs
@@ -93,10 +93,10 @@ mod usable_size {
 			pub unsafe extern "C" fn malloc_usable_size(ptr: *const c_void) -> usize {
 				// mimalloc doesn't actually mutate the value ptr points to,
 				// but requires a mut pointer in the API
-				mimalloc_sys::mi_usable_size(ptr as *mut _)
+				libmimalloc_sys::mi_usable_size(ptr as *mut _)
 			}
 
-		} else if #[cfg(target_os = "linux")] {
+		} else if #[cfg(any(target_os = "linux", target_os = "android"))] {
 
 			/// Linux call system allocator (currently malloc).
 			extern "C" {